# tools/api-docs/public/openapi.yaml
openapi: 3.0.3
info:
  title: Business Data Streaming Pipeline — HTTP APIs
  description: |
    Tài liệu tương tác cho demo k3s. **Chọn đúng Server** (góc trên) trước Try it out.

    Request đi qua **Vite proxy** (cùng origin `:5190`, tránh CORS). Chọn Server:

    | Server | Backend thật |
    |--------|----------------|
    | `/proxy/dashboard` | dashboard-api :32000 |
    | `/proxy/tracking` | tracking-api :31000 |
    | `/proxy/commerce` | commerce-backend :30330 |

    Backend trên WSL IP: tạo `tools/api-docs/.env` với `VITE_PROXY_DASHBOARD=http://<WSL_IP>:32000` (và tương tự tracking/commerce), restart `npm run dev`.

    Tham số kỳ: `minutes` (rolling) hoặc `date=YYYY-MM-DD` (UTC, ưu tiên hơn minutes).
  version: 1.0.0
servers:
  - url: /proxy/dashboard
    description: dashboard-api — qua Vite proxy → :32000
  - url: /proxy/tracking
    description: tracking-api — qua Vite proxy → :31000
  - url: /proxy/commerce
    description: commerce-backend — qua Vite proxy → :30330
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Dashboard login — POST /api/auth/login
    ingestKey:
      type: http
      scheme: bearer
      description: TRACKING_INGEST_API_KEY (Lap2 adapter)
  parameters:
...
    limit:
      name: limit
      in: query
      schema:
        type: integer
        example: 20
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
...
paths:
...
  /api/auth/login:
    post:
      tags: [dashboard-auth]
      servers:
        - url: /proxy/dashboard
      summary: Đăng nhập dashboard
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200": { description: JWT + user }
        "401": { description: invalid_credentials }
  /api/auth/me:
    get:
      tags: [dashboard-auth]
      servers: [{ url: /proxy/dashboard }]
      summary: Thông tin user hiện tại
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: user profile }
        "401": { description: unauthorized }
  /api/auth/change-password:
    patch:
      tags: [dashboard-auth]
      servers: [{ url: /proxy/dashboard }]
      summary: Đổi mật khẩu
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string }
                new_password: { type: string, minLength: 6 }
      responses:
        "200": { description: ok }
...
  /api/system/pipeline:
    get:
      tags: [dashboard-admin]
      servers: [{ url: /proxy/dashboard }]
      summary: Admin — trạng thái pipeline
      security: [{ bearerAuth: [] }]
      description: Role super_admin
      responses:
        "200": { description: health + metrics }
  /api/system/setup:
    get:
      tags: [dashboard-admin]
      servers: [{ url: /proxy/dashboard }]
      summary: Admin — gợi ý port/env demo
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: setup hints }
  /api/chat/insights:
    get:
      tags: [dashboard-admin]
      servers: [{ url: /proxy/dashboard }]
      summary: Admin — insight Qdrant gần đây
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/limit"
      responses:
        "200": { description: insights }
...
  /api/orders:
    post:
      tags: [commerce]
      servers: [{ url: /proxy/commerce }]
      summary: Tạo đơn (web-demo) → RabbitMQ
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [anonymousId, sessionId, items]
              properties:
                anonymousId: { type: string }
                sessionId: { type: string }
                items:
                  type: array
                  items: { type: object }
      responses:
        "201": { description: orderCode }
  /api/orders/{orderCode}/cancel:
    post:
      tags: [commerce]
      servers: [{ url: /proxy/commerce }]
      summary: Hủy đơn
      parameters:
        - name: orderCode
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: cancelled }
  /commerce/add-to-cart:
    post:
      tags: [commerce]
      servers: [{ url: /proxy/commerce }]
      summary: Commerce event — add to cart
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [anonymous_id, session_id, product_id]
              properties:
                anonymous_id: { type: string }
                session_id: { type: string }
                product_id: { type: string }
      responses:
        "202": { description: accepted }
  /commerce/checkout-start:
    post:
      tags: [commerce]
      servers: [{ url: /proxy/commerce }]
      summary: Commerce event — checkout start
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [anonymous_id, session_id]
      responses:
        "202": { description: accepted }
  /commerce/purchase:
    post:
      tags: [commerce]
      servers: [{ url: /proxy/commerce }]
      summary: Commerce event — purchase succeeded
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [anonymous_id, session_id]
      responses:
        "202": { description: accepted }
  /commerce/payment-failed:
    post:
      tags: [commerce]
      servers: [{ url: /proxy/commerce }]
      summary: Commerce event — payment failed
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [anonymous_id, session_id]
      responses:
        "202": { description: accepted }
  /commerce/cart-abandoned:
    post:
      tags: [commerce]
      servers: [{ url: /proxy/commerce }]
      summary: Commerce event — cart abandoned
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [anonymous_id, session_id]
      responses:
        "202": { description: accepted }